requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
chromadb==0.5.0
requests==2.32.4
openai==1.98.0
python-dotenv==1.1.1
pymupdf
